* Changes to status codes     - removed the VAL_AC_TRUSTED_ZONE and VAL_TRUSTED_ZONE codes     - renamed *PROVABLY_INSECURE* to *PINSECURE*     - renamed VAL_LOCAL_ANSWER to VAL_OOB_ANSWER * Changes to types     - using u_char in place of u_int8_t to represent character buffers     - using size_t where appropriate     - Decoupled val_rrset_rec from other internal structures. Moved members     from older val_rrset_digested structure to rrset_rec structure     - No longer store complete header information in val_rrset_rec; only keep     rcode information * Changes to function prototypes     - using higher order types instead of u_int* variants for exported     functions     - val_add_valpolicy() now takes a void * as the second argument     - No longer pass OTW format domain names for val_resolve_and_check() * policy-related     - No longer define a "trusted" zone security expectation since there is no     way to give it special treatment in the high-level API; the "ignore"     condition subsumes the "trusted" condition.     - Define new libval_policy_definition_t structure to encapsulate the     libval-specific policy definition parameters in val_add_valpolicy()

+++ dnssec-tools/validator/apps/getrrset.c
@@ -100,7 +100,7 @@ print_results(struct val_answer_chain *results)
         
         for (j=1, rr = res->val_ans; rr; rr=rr->rr_next) {
             char            buf[1028];
-            int             buflen = 1024;
+            int          buflen = 1024;
             
             get_hex_string(rr->rr_data, rr->rr_length, buf, buflen);
             fprintf(stderr, "RR %d : %s\n", j++, buf); 
@@ -118,7 +118,7 @@ main(int argc, char *argv[])
     char           *node = NULL;
     int             retval;
     val_log_t      *logp;
-    u_int16_t      type_h = ns_t_a;
+    int      type_h = ns_t_a;
     struct val_answer_chain *results = NULL;
     int success = 0;
 
@@ -153,7 +153,7 @@ main(int argc, char *argv[])
             break;
 
         case 't':
-            type_h = res_nametotype(optarg, &success);
+            type_h = (int)res_nametotype(optarg, &success);
             if (!success) {
                 fprintf(stderr, "Unrecognized type %s\n", optarg);
                 usage (argv[0]);
